Move the libanki.Utils.writeToFile() method to ResourceLoader
This method was used in two places:
- Shared.getTestFilePath() which was an unused method(so it was deleted)
- ResourceLoader

Note that writeToFile called writeToFileImpl which was moved with
it(no other usages).
